refactor(matrix): extract reaction cluster into MatrixReactionsMixin (god-file Phase 3)
MatrixAdapter is 3,771 lines and 92 methods. This lifts its reaction cluster
into plugins/platforms/matrix/reactions_mixin.py, the same mechanical mixin
extraction that produced gateway/authz_mixin.py and the SessionDB mixins in
hermes_state_*.py.

Matrix has no native buttons, so reactions are the interaction surface: they
signal that a turn started or finished, and they are how a user answers an
approval prompt or a model picker. The adapter already fenced this block off
under its own banner, 'Reactions (send, receive, processing lifecycle)'; this
takes exactly that block.

Moved verbatim (350 lines, 13 methods):

  _send_reaction                  _validate_matrix_prompt_reactor
  _redact_reaction                _send_invalid_reaction_feedback
  _schedule_reaction_redaction    _expire_matrix_approval_prompt
  on_processing_start             _expire_matrix_model_picker_prompt
  on_processing_complete          _redact_bot_approval_reactions
  _on_reaction                    _redact_bot_model_picker_reactions
  _matrix_prompt_expired

Behavior-neutral. Every body is byte-identical to what it replaced; the only
edits are the class declaration, one import, and the new module header.

Base order matters here and is deliberate: on_processing_start and
on_processing_complete override BasePlatformAdapter methods, so
MatrixReactionsMixin precedes it in the bases and wins exactly as the
adapter's own definitions did.

logger is bound by explicit name, so records keep the name
plugins.platforms.matrix.adapter and getLogger returns the adapter's own
object. EventType and RoomID are imported under the adapter's own ImportError
guard; the moved code uses RoomID only as a cast and EventType only for the
REACTION string constant, so the real mautrix types and the fallback stubs
behave alike.

adapter.py 4,999 -> 4,633 lines; the class 3,771 -> 3,404 lines and 92 -> 79 methods.
